1. 重构pom.xml依赖关系 2. 补充log模块业务实现

+++ adam-manage-api/src/main/java/com/pilot/entity/param/user/UserDeleteParam.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Data
 public class UserDeleteParam {
     // 删除用户id列表
-    private List<Integer> userIds;
+    private List<Long> userIds;
 }
